// src/api-client/models/Note.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Quatt API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.2
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from "../runtime";
/**
 *
 * @export
 * @interface Note
 */
export interface Note {
  /**
   *
   * @type {string}
   * @memberof Note
   */
  id: string;
  /**
   *
   * @type {string}
   * @memberof Note
   */
  note: string;
  /**
   *
   * @type {Date}
   * @memberof Note
   */
  createdAt?: Date;
  /**
   *
   * @type {string}
   * @memberof Note
   */
  updatedBy?: string | null;
}

/**
 * Check if a given object implements the Note interface.
 */
export function instanceOfNote(value: object): boolean {
  let isInstance = true;
  isInstance = isInstance && "id" in value;
  isInstance = isInstance && "note" in value;

  return isInstance;
}

export function NoteFromJSON(json: any): Note {
  return NoteFromJSONTyped(json, false);
}

export function NoteFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): Note {
  if (json === undefined || json === null) {
    return json;
  }
  return {
    id: json["id"],
    note: json["note"],
    createdAt: !exists(json, "createdAt")
      ? undefined
      : new Date(json["createdAt"]),
    updatedBy: !exists(json, "updatedBy") ? undefined : json["updatedBy"],
  };
}

export function NoteToJSON(value?: Note | null): any {
  if (value === undefined) {
    return undefined;
  }
  if (value === null) {
    return null;
  }
  return {
    id: value.id,
    note: value.note,
    createdAt:
      value.createdAt === undefined ? undefined : value.createdAt.toISOString(),
    updatedBy: value.updatedBy,
  };
}