// src/api-client/models/AdminAddInstallationNote200Response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Quatt API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.2
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from "../runtime";
import type { Note } from "./Note";
import { NoteFromJSON, NoteFromJSONTyped, NoteToJSON } from "./Note";
import type { ResponseMeta } from "./ResponseMeta";
import {
  ResponseMetaFromJSON,
  ResponseMetaFromJSONTyped,
  ResponseMetaToJSON,
} from "./ResponseMeta";

/**
 *
 * @export
 * @interface AdminAddInstallationNote200Response
 */
export interface AdminAddInstallationNote200Response {
  /**
   *
   * @type {ResponseMeta}
   * @memberof AdminAddInstallationNote200Response
   */
  meta: ResponseMeta;
  /**
   *
   * @type {Note}
   * @memberof AdminAddInstallationNote200Response
   */
  result: Note;
}

/**
 * Check if a given object implements the AdminAddInstallationNote200Response interface.
 */
export function instanceOfAdminAddInstallationNote200Response(
  value: object,
): boolean {
  let isInstance = true;
  isInstance = isInstance && "meta" in value;
  isInstance = isInstance && "result" in value;

  return isInstance;
}

export function AdminAddInstallationNote200ResponseFromJSON(
  json: any,
): AdminAddInstallationNote200Response {
  return AdminAddInstallationNote200ResponseFromJSONTyped(json, false);
}

export function AdminAddInstallationNote200ResponseFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): AdminAddInstallationNote200Response {
  if (json === undefined || json === null) {
    return json;
  }
  return {
    meta: ResponseMetaFromJSON(json["meta"]),
    result: NoteFromJSON(json["result"]),
  };
}

export function AdminAddInstallationNote200ResponseToJSON(
  value?: AdminAddInstallationNote200Response | null,
): any {
  if (value === undefined) {
    return undefined;
  }
  if (value === null) {
    return null;
  }
  return {
    meta: ResponseMetaToJSON(value.meta),
    result: NoteToJSON(value.result),
  };
}